Increase time to wait for Google Meets status screen
The status screen was taking longer than 600ms to load,
which meant that the Puppeteer browser bot wasn't waiting long
enough to see the Meet status. Increasing the timeout is still
hacky but as long as there is no reliable Meets API, we are
likely stuck with this method.

Change-type: patch
Signed-off-by: Christina Ying Wang <christina@balena.io>